drone system mock-up
The system shows the picture of a map in fullscreen.
The user can create as many drones as he/she wants to
When the system loads, some accidents are loaded with it. This means that with PHP you create some accidents that are displayed in the page as soon as the user loads it
The user can select a drone (previously created) and tell it to fly to the spot where the accident is displayed
The system creates statistics. The system displays how many times the drone has been used to attend an accident. The system shows how many drones you have in total
The user can destroy/delete/remove drones as he/she wants it
No need to use a database at all. When the user refreshes the page, the system shows new random accidents and the user is able to start creating drones and command them to fly to any accident location
Questions and details will be discussed on Friday the 4th of March during class